refactor(community): move components to a shared directory structure
