+++ apps/community/src/app/about/dept/page.tsx
@@ -1,7 +1,7 @@
 import { Fragment } from 'react';
 import { sprinkles } from '@aics-client/design-system/styles';
-import { List } from '~/app/about/components/list';
-import { Section } from '~/app/about/components/section';
+import { List } from '~/components/about/list';
+import { Section } from '~/components/about/section';
 import { getDepts } from '~/app/about/dept/remotes';
 import { PageHeader } from '~/components/page-header';
