+++ apps/community/src/components/about/list.css.ts
