+++ apps/community/src/components/about/section.css.ts
