+++ apps/community/src/components/lab/lab-card.css.ts
